repair ETHERNET and PCIEC

+++ board-salvator/PCIEC/012_pciec_tranfers_data_from_host_pc_to_board.sh
@@ -6,10 +6,16 @@ set -e
 
 echo "\n*****************TRANFERS DATA FROM HOST PC TO BOARD******************\n"
 
-eval $MOUNT_RAM
+mkdir -p $RAM_DIR
+
+# Mount ram
+$(dirname $0)/../common/mount-device.sh $RAM_DIR
 
 for size in "1" "50" "350"; do
 
+	rm -rf $RAM_DIR/*
+
+	echo "prepare data for test"
 
 #check file on host pc
 ftp -inv $IPSERVER > $LOGFILE 2>&1 <<END_SCRIPT
@@ -25,20 +31,28 @@ END_SCRIPT
 #---------
 
 	if grep -i "file-${size}mb" $LOGFILE >/dev/null;then
-		echo "prepare data successfully"
+		echo "prepare data file-${size}mb successfully"
 	else
-		echo "prepare data not successfully"
+		echo "prepare data file-${size}mb not successfully please create data on host pc"
+		rm -rf $LOGFILE
 		eval $FAIL_MEG
-		break;	
+		exit 1;	
 	fi
 
-	rm -rf $LOGFILE
+	if [ -f $LOGFILE ];then
+	    rm -rf $LOGFILE
+	fi 
 
 	echo "tranfer file ${size}MB"
 	$(dirname $0)/ftp_get_pc_to_board_data.sh $size
 
 done
 
-eval $UNMOUNT_RAM
+sync
+
+# Umount ram
+$(dirname $0)/../common/umount-device.sh $RAM_DIR
+
+rm -rf $RAM_DIR
 
 echo "\n*********************************************************************\n"
